Math Worksheet Gen project/requirements.txt

Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Login>=0.5
Flask-Cors # For Cross-Origin Resource Sharing
Werkzeug>=2.0 # Flask dependency for password hashing

AI Interaction

google-generativeai

File Generation

fpdf2 # For PDF generation
python-docx # For DOCX generation
pdf2image # Optional: For PNG/JPG generation (requires external Poppler)

Authentication

Authlib>=1.0 # For OAuth (Google Login)